NeuroOps/src/visual_cortex/reid.py
import torch
import numpy as np
import cv2
from PIL import Image
# Ideally import fastreid, but for prototype we can use a simple ResNet or CLIP
# We will reuse CLIP here as a baseline 'Identity' embedder if specific reid model isn't downloaded,
# or we can mock it.
# Let's assume we want to use CLIP's visual encoder for re-id as a strong baseline simplifier.

from sentence_transformers import SentenceTransformer

class IdentityEncoder:
    def __init__(self, model_name="clip-ViT-B-32"):
        # Utilizing CLIP for Re-ID as a "Zero-Shot Identity" baseline
        # In production: replace with fastreid.modeling.build_model
        try:
            self.model = SentenceTransformer(model_name)
            print("[RE-ID] Loaded Identity Encoder (CLIP baseline)")
        except Exception as e:
            print(f"[RE-ID] Error loading model: {e}")
            self.model = None

    def extract_feature(self, image_crop):
        """
        Extract feature vector for a person crop.
        """
        if self.model is None:
            return np.zeros(512).tolist()
            
        if isinstance(image_crop, np.ndarray):
            image = Image.fromarray(image_crop)
        else:
            image = image_crop
            
        vector = self.model.encode(image)
        return vector.tolist()
